reworked runner
now this is awkward: First of all, we need max two runners: All worlds should be simultanously updated. Additionally, spawning subprocesses triggers a copy-on-write the first time one of the procceses changes the arguments.
Thus, threads or processes don't seem feasable at the moment.
Kind-of hacky intermediate solution: spawn a subprocess that continously triggers the step() function via http. the step will then compute when to step nodenets and worlds
we need to remove the http-overhead, and we need to make the logilfes readable again. Or, we need to come up with more elaborate solution.