node.activation now is a dynamic parameter:
the activation of a node is determined by its 'gen' slot activation (or more genral: the sum of its slot activations). this is on the fly calculated. Now, changing the activation of a node changes
the 'gen' slot activation. previously, manual changes to node.activation where ignored, since the node's activation was recalculated from the slots anyhow, so this property never served any purpose