discover worlds during runtime init
this gets rid of the stupid import-section in world.py.
World implementations however are supposed to say in their __init__ which modules contain world classes.

using __all__ for this is kind of a hack, because the world might need additional things to work (like, worldadapters)
currently the world needs to import them itself.
Otoh, we're looking for worlds we need to know during runtime init, not for their WAs... maybe revisit this.